refactor(tui): move renderGroupedRootTools into the grouped-root module
Pure move, no behavior change. PR #773 extracted `formatGroupedToolResults`
into tool-lane-render-grouped-root.ts but left its only caller behind, so the
grouped-root concern spanned two files while tool-lane-render-agent.ts carried
both the agent-frame and grouped-root renderers.

Moving the caller to join its helper puts the whole concern in one module and
takes tool-lane-render-agent.ts from 338 to 264 lines, which is what makes room
for the head-row clamp in the next commit without crossing the 350-line ceiling.

`clampLineToTerminal` leaves render-agent's imports because the moved function
was its only user there; the next commit reintroduces it for the head rows.
Re-exports move with the symbols, so tool-lane.ts's import surface is unchanged.
